+++ test/programs/fql/ssh-simplified/s3_clnt_1.cil.c
@@ -108,7 +108,7 @@ int ssl3_connect(int initial_state )
 #line 61
   blastFlag = 0;
 #line 62
-  tmp = __BLAST_NONDET;
+  tmp = (unsigned long ) __BLAST_NONDET;
 	s__info_callback = __BLAST_NONDET;
 	s__ctx__info_callback = __BLAST_NONDET;
 	s__version = __BLAST_NONDET;
